Iceberg sink! Can append to a table, with test.
Splits sink tests into its own package.
